requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
alembic==1.14.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
base58==2.1.1
bcrypt==4.2.1
billiard==4.2.1
build==1.2.2.post1
CacheControl==0.14.2
cachetools==5.5.2
celery==5.4.0
cement==2.10.14
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coincurve==21.0.0
colorama==0.4.6
contourpy==1.3.1
Counter==1.0.0
crispy-bootstrap5==2025.4
cryptography==44.0.0
cycler==0.12.1
cytoolz==1.0.1
Django==5.1.6
django-crispy-forms==2.4
django-taggit==6.1.0
django-widget-tweaks==1.5.0
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
eth-hash==0.7.1
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
exceptiongroup==1.2.2
Faker==37.1.0
fastapi==0.115.12
fastapi-users==14.0.1
fastapi-users-db-sqlalchemy==7.0.0
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-oauth==0.16.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.6
jmespath==1.0.1
jose==1.0.0
kiwisolver==1.4.8
kombu==5.4.2
litestar==2.15.2
litestar-htmx==0.4.1
makefun==1.15.6
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
msgpack==1.1.0
msgspec==0.19.0
multidict==6.2.0
network==0.1
networkx==3.4.2
nodeenv==1.9.1
numpy==2.2.4
opencv-python==4.11.0.86
orjson==3.10.15
packaging==24.2
parsimonious==0.10.0
passlib==1.7.4
pathspec==0.10.1
pillow==11.1.0
pip-tools==7.4.1
pluggy==1.5.0
polyfactory==2.21.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pwdlib==0.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
# pypiwin32==223
pyproject_hooks==1.2.0
pyright==1.1.393
pytest==8.3.5
pytest-asyncio==0.25.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
redis==6.1.0
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rich-click==1.8.8
rsa==4.9
s3transfer==0.11.4
semantic-version==2.10.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
starlette==0.45.3
termcolor==2.5.0
tomli==2.2.1
toolz==1.0.0
tronpy==0.5.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.3
pytz==2024.1
djangorestframework==3.16.0